+++ src_new/scripts/levi_ataxia/levi/levi_scripts/shaman/004_DAMAGE_PRIOS.lua
@@ -33,7 +33,7 @@ if checkAffList({"anorexia", "asthma", "slickness", "bloodfire"},3) then
 
 	end
   
-if checkAffList({"healthleech", "asthma", "weariness", "sensitivity", "clumsiness", "hypochondria", "parasite", "rebbies"},3) then
+if checkAffList({"healthleech", "asthma", "weariness", "sensitivity", "clumsiness", "parasite", "rebbies"},3) then
 		kelpstack = true
   else
   kelpstack = false